@@ -2,6 +2,7 @@
   "name": "ydimage-server",
   "version": "1.0.0",
   "description": "",
+  "type": "module",
   "main": "index.js",
   "scripts": {
     "start": "node index.js"
@@ -9,15 +10,16 @@
   "author": "Yordan Doykov",
   "license": "ISC",
   "dependencies": {
-    "aws-sdk": "^2.1171.0",
-    "body-parser": "^1.20.0",
+    "@aws-sdk/client-s3": "^3.462.0",
+    "@aws-sdk/lib-storage": "^3.462.0",
+    "body-parser": "^1.20.2",
     "cors": "^2.8.5",
-    "dotenv": "^16.0.3",
-    "express": "^4.18.1",
+    "dotenv": "^16.3.1",
+    "express": "^4.18.2",
     "multer": "^1.4.5-lts.1",
     "path": "^0.12.7"
   },
   "engines": {
-    "node": "18.x"
+    "node": "20.x"
   }
 }
